refactor: Rename concurrentStepGroupUtils -> stepHierarchyUtils to match the StepHierarchy type.
